Combine path conversion functions
