+++ tests/test_utils.py
@@ -1,6 +1,9 @@
 # Copyright (C) 2012 Anaconda, Inc
 # SPDX-License-Identifier: BSD-3-Clause
+from __future__ import annotations
+
 import sys
+from contextlib import nullcontext
 from logging import getLogger
 from os import environ, pathsep
 from os.path import dirname, join
@@ -239,3 +242,23 @@ def test_ensure_dir_errors():
             get_test_dir()
 
     assert exc_message in str(exc_info.value)
+
+
+@pytest.mark.parametrize(
+    "function,deprecated,raises",
+    [
+        ("path_identity", False, TypeError),
+        ("unix_path_to_win", False, TypeError),
+        ("win_path_to_cygwin", False, TypeError),
+        ("cygwin_path_to_win", False, TypeError),
+    ],
+)
+def test_deprecations(
+    function: str,
+    deprecated: bool,
+    raises: type[Exception] | None,
+) -> None:
+    deprecated_context = pytest.deprecated_call() if deprecated else nullcontext()
+    raises_context = pytest.raises(raises) if raises else nullcontext()
+    with deprecated_context, raises_context:
+        getattr(utils, function)()
